requirement.txt
django==1.11
djangorestframework
django-categories
mysqlclient
Pillow
facepy
